requirements.txt
qdrant-client
openai
python-dotenv
fastapi
uvicorn
pydantic==2.*
tiktoken
